requirements.txt
#torch
#torchvision
numpy #1.19.5
# matplotlib
keras-facenet
#deep-sort-realtime
pillow
pyqt5
opencv-python
# mtcnn
imutils
tensorflow-gpu =='2.5.0'
#scipy=='1.9.2'
#tensorflow < '2.11'
#python -c "import tensorflow as tf; print(tf.config.list_physical_devices('GPU'))"
#pip install tensorflow==2.10.0
retina-face